+++ lib/puppet/provider/st2_pack/default.rb
@@ -10,8 +10,7 @@ Puppet::Type.type(:st2_pack).provide(:default) do
   def st2_authenticate
     # Reuse previous token
     return @token if @token
-    @token = exec_st2('auth', resource[:user], '-t', '-p', resource[:password],
-                      sensitive: true).chomp
+    @token = exec_st2('auth', resource[:user], '-t', '-p', resource[:password]).chomp
   end
 
   def create
@@ -50,20 +49,26 @@ Puppet::Type.type(:st2_pack).provide(:default) do
 
   # execute the st2 command and use the system locale (UTF8)
   # so that the st2 CLI doesn't complain and throw errors
-  def exec_st2(*args, sensitive: false)
+  def exec_st2(*args)
     # escape all arguments so they're safe to use in a shell command
     escaped_args = args.map { |a| Shellwords.shellescape(a) }
+
+    # when passing in a command array into Puppet::Util::Execution.execute()
+    # it doesn't do shell expansion on the arguments, but special characters are still
+    # sometimes not processed correctly by the underlying system.
+    # instead if you specify the first argument as a string, the command is treated
+    # as a shell command with normal shell expansion rules and our escaping above
+    # works correctly
+    # see: https://ruby-doc.org/core/Kernel.html#method-i-exec
+    command_str = ([command(:st2)] + escaped_args).join(' ')
+
     # when we started passing in the override_locale: option, there is some "known behavior"
     # of this function where when any option is passed in it sets failonfail: false and
     # combine: false for some terrible reason. We want both of those set to true like
     # they are when no options are specified, so we set them explicitly.
-    #
-    # We also have the option to mark the command as sensitive for when we auth.
-    # This prevents passwords from being written to the log.
-    Puppet::Util::Execution.execute([command(:st2)] + escaped_args,
+    Puppet::Util::Execution.execute(command_str,
                                     override_locale: false,
                                     failonfail: true,
-                                    combine: true,
-                                    sensitive: sensitive)
+                                    combine: true)
   end
 end